+++ src/main/java/net/fosforito/partido/model/split/SplitDTO.java
@@ -10,14 +10,14 @@ public class SplitDTO {
   private double paid;
 
   @NotNull
-  private int partsOfBill;
+  private double partsOfBill;
 
   public SplitDTO() {
   }
 
   public SplitDTO(Long debtor,
                   @NotNull double paid,
-                  @NotNull int partsOfBill) {
+                  @NotNull double partsOfBill) {
     this.debtor = debtor;
     this.paid = paid;
     this.partsOfBill = partsOfBill;
@@ -39,11 +39,11 @@ public class SplitDTO {
     this.paid = paid;
   }
 
-  public int getPartsOfBill() {
+  public double getPartsOfBill() {
     return partsOfBill;
   }
 
-  public void setPartsOfBill(int partsOfBill) {
+  public void setPartsOfBill(double partsOfBill) {
     this.partsOfBill = partsOfBill;
   }
 }
